backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
easyocr
motor
pydantic
passlib[bcrypt]
pyjwt
rapidfuzz
numpy
opencv-python-headless
python-dotenv
reportlab
pymupdf
tensorflow
torch
torchvision
torch_geometric